requirements.txt
streamlit~=1.38.0
openai~=1.50.0
python_dotenv~=1.0.0
opencv-python~=4.10.0.84
easyocr~=1.7.2
